+++ base/include/asm-x86_64/rtai_srq.h
@@ -30,7 +30,7 @@ static inline long long rtai_srq(long srq, long args)
 	return retval;
 }
 
-static inline int rtai_open_srq(unsigned int label)
+static inline int rtai_open_srq(unsigned long label)
 {
 	return (int)rtai_srq(0, label);
 }
